move header.js and add some styles
