+++ src/components/header/header.css
@@ -1,0 +1,23 @@
+.header {
+  padding: 1rem 0;
+  margin-bottom: 3rem;
+  border-bottom: 1px solid var(--color-light);
+}
+.header > .container-md {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+.header .website-title {
+  text-align: center;
+  margin: 0;
+}
+.header .website-title .link-logo {
+  font-size: 2.2rem;
+  font-family: 'Circular Std Bold';
+  color: var(--color-primary);
+  display: inline-block;
+}
+.header .website-title .link-logo:hover {
+  color: var(--color-secondary);
+}
