+++ src/components/header/header.js
@@ -1,0 +1,74 @@
+import { Link } from 'gatsby'
+import PropTypes from 'prop-types'
+import React from 'react'
+
+import './header.css'
+
+const Header = ({ siteTitle }) => (
+  <header class="header">
+    <div class="container-md">
+      <h1 class="website-title">
+        <Link
+          to="/old"
+          class="link-logo"
+          aria-label="omar houmz"
+          role="heading"
+          aria-level="1"
+        >
+          oh
+        </Link>
+      </h1>
+      <nav class="navigation dark">
+        <ul>
+          <li>
+            <Link to="/old" class="active">
+              Home
+            </Link>
+          </li>
+          <li>
+            <Link to="./about">About</Link>
+          </li>
+          <li>
+            <Link to="./projects">Projects</Link>
+          </li>
+        </ul>
+      </nav>
+    </div>
+  </header>
+  // <header
+  //   style={{
+  //     background: `rebeccapurple`,
+  //     marginBottom: `1.45rem`,
+  //   }}
+  // >
+  //   <div
+  //     style={{
+  //       margin: `0 auto`,
+  //       maxWidth: 960,
+  //       padding: `1.45rem 1.0875rem`,
+  //     }}
+  //   >
+  //     <h1 style={{ margin: 0 }}>
+  //       <Link
+  //         to="/"
+  //         style={{
+  //           color: `white`,
+  //           textDecoration: `none`,
+  //         }}
+  //       >
+  //         {siteTitle}
+  //       </Link>
+  //     </h1>
+  //   </div>
+  // </header>
+)
+
+Header.propTypes = {
+  siteTitle: PropTypes.string,
+}
+
+Header.defaultProps = {
+  siteTitle: ``,
+}
+
+export default Header
